Develop

+++ Core/RggFachwerk.pas
@@ -23,10 +23,10 @@ uses
 
 type
   TGeometrie = array [0 .. 1, 1 .. 9] of Integer;
-  TKnotenVektor = array [1 .. 6] of double;
-  TStabVektor = array [1 .. 9] of double;
+  TKnotenVektor = array [1 .. 6] of single;
+  TStabVektor = array [1 .. 9] of single;
   Lagerkraefte = (AX, AY, BX, BY);
-  TAufLager = array [Lagerkraefte] of double;
+  TAufLager = array [Lagerkraefte] of single;
 
 const
   { Geometriematrix, Stab in Spalte verbindet die beiden Knoten }
@@ -59,7 +59,7 @@ type
     K: Integer; { Anzahl der Knoten }
     K1: Integer; { KnotenNr. des Festlagers A }
     K2: Integer; { KnotenNr. des Loslagers B }
-    Phi: double; { Winkel von FB in Grad }
+    Phi: single; { Winkel von FB in Grad }
     S: Integer; { Anzahl der Staebe }
     G: TGeometrie; { Geometriematrix }
     vektorEA: TStabVektor; { Vektor EA }
@@ -68,19 +68,19 @@ type
     FXsaved, FYsaved: TKnotenVektor; { Kopie des Belastungsvektors }
     FO, FO1, FO2: TKnotenVektor; { Verschiebungen }
     Lager: TAufLager; { Auflagerreaktionen }
-    P: double; { Winkel der Auflagekraft im Loslager B in Rad }
+    P: single; { Winkel der Auflagekraft im Loslager B in Rad }
     Q: TStabVektor; { l/EA Vektor }
     FS1: TStabVektor; { Puffer für Stabkräfte }
     FS: TStabVektor; { Speicher-Vektor der Stabkräfte bei äußerer Last }
-    SummeFX, SummeFY, SummeMO: double; { summierte Belastungen }
-    DX1, DY1, W1: double;
-    DX2, DY2, W2: double;
-    DX3, DY3, W3: double;
-    DX4, DY4, W4: double;
-    D, D1, D2, W: double;
-    BekanntFX, BekanntFY: double;
-    WantenPower: double;
-    MastDruck: double;
+    SummeFX, SummeFY, SummeMO: single; { summierte Belastungen }
+    DX1, DY1, W1: single;
+    DX2, DY2, W2: single;
+    DX3, DY3, W3: single;
+    DX4, DY4, W4: single;
+    D, D1, D2, W: single;
+    BekanntFX, BekanntFY: single;
+    WantenPower: single;
+    MastDruck: single;
     SalingTyp: TSalingTyp;
     BerechneVerschiebungen: Boolean;
     constructor Create;
@@ -89,7 +89,7 @@ type
     procedure KG21(l, l1, l2, l3, i1, i2, i3: Integer);
     procedure Stabkraefte;
     procedure StabkraefteOSB;
-    procedure Auflagerkraefte(SumFX, SumFY, SumMO: double; out Lager: TAuflager);
+    procedure Auflagerkraefte(SumFX, SumFY, SumMO: single; out Lager: TAuflager);
     procedure Verschiebungen;
     procedure ActionF;
   end;
@@ -243,9 +243,9 @@ begin
   KG21(6, 5, 4, 2, 9, 7, 3);
 end;
 
-procedure TFachwerk.Auflagerkraefte(SumFX, SumFY, SumMO: double; out Lager: TAuflager);
+procedure TFachwerk.Auflagerkraefte(SumFX, SumFY, SumMO: single; out Lager: TAuflager);
 var
-  FB: double;
+  FB: single;
 begin
   FB := -SumMO / ((KX[K2] - KX[K1]) * sin(P) - (KY[K2] - KY[K1]) * cos(P));
   Lager[BX] := FB * cos(P);
@@ -257,10 +257,10 @@ end;
 procedure TFachwerk.Verschiebungen;
 var
   i, j, l: Integer;
-  Sum1FX, Sum1FY, Sum1MO: double;
+  Sum1FX, Sum1FY, Sum1MO: single;
   Lager1: TAufLager;
-  PORad: double;
-  F: double;
+  PORad: single;
+  F: single;
 begin
   PORad := 0;
   for l := 1 To K do
