+++ Graph/RggVector.pas
@@ -4,53 +4,53 @@ interface
 
 type
   vec2 = record
-    x, y, z: double;
+    x, y, z: single;
   end;
 
   vec3 = record
-    x, y, z: double;
+    x, y, z: single;
   end;
 
-function InitPoint2D(ax: double; ay: double): vec2;
-function Mag2D(v: vec2): double;
+function InitPoint2D(ax: single; ay: single): vec2;
+function Mag2D(v: vec2): single;
 function Normalize2D(v: vec2): vec2;
 
-function InitPoint3D(ax: double; ay: double; az: double): vec3;
-function Mag3D(v: vec3): double;
+function InitPoint3D(ax: single; ay: single; az: single): vec3;
+function Mag3D(v: vec3): single;
 function Normalize3D(v: vec3): vec3;
 function Subtract(v1, v2: vec3): vec3;
 function Cross(v1, v2: vec3): vec3;
-function Divide(v :vec3; num: double): vec3;
-function Dot(v1, v2: vec3): double;
+function Divide(v :vec3; num: single): vec3;
+function Dot(v1, v2: vec3): single;
 
-function LookUpRa10(Index: Integer): double;
+function LookUpRa10(Index: Integer): single;
 
 const
   vec3Null: vec3 = (x: 0; y: 0; z: 0);
 
 implementation
 
-function InitPoint2D(ax: double; ay: double): vec2;
+function InitPoint2D(ax: single; ay: single): vec2;
 begin
   result.x := ax;
   result.y := ay;
 end;
 
-function InitPoint3D(ax: double; ay: double; az: double): vec3;
+function InitPoint3D(ax: single; ay: single; az: single): vec3;
 begin
   result.x := ax;
   result.y := ay;
   result.z := az;
 end;
 
-function Mag2D(v: vec2): double;
+function Mag2D(v: vec2): single;
 begin
   result := sqrt(v.x * v.x + v.y * v.y);
 end;
 
 function Normalize2D(v: vec2): vec2;
 var
-  d: double;
+  d: single;
 begin
   result := InitPoint2D(v.x, v.y);
   d := Mag2D(result);
@@ -62,7 +62,7 @@ begin
 end;
 
 { Calculate the magnitude of the vector }
-function Mag3D(v: vec3): double;
+function Mag3D(v: vec3): single;
 begin
   result := sqrt(v.x * v.x + v.y * v.y + v.z * v.z);
 end;
@@ -90,7 +90,7 @@ begin
 end;
 
 { Divide the scalar number into the vector v }
-function Divide(v: vec3; num: double): vec3;
+function Divide(v: vec3; num: single): vec3;
 var
   d: vec3;
 begin
@@ -106,7 +106,7 @@ end;
 { Normalize the vector v }
 procedure Normalize(var v: vec3);
 var
-  d: double;
+  d: single;
 begin
   d := sqrt(v.x * v.x + v.y * v.y + v.z * v.z);
   if (d <> 0) then
@@ -119,7 +119,7 @@ end;
 
 function Normalize3D(v: vec3): vec3;
 var
-  d: double;
+  d: single;
 begin
   result := InitPoint3D(v.x, v.y, v.z);
   d := Mag3D(result);
@@ -132,14 +132,14 @@ begin
 end;
 
 { Calculate the dot product of the two vectors v1 and v2 }
-function Dot(v1, v2: vec3): double;
+function Dot(v1, v2: vec3): single;
 begin
   result := v1.x * v2.x + v1.y * v2.y + v1.z * v2.z;
 end;
 
-function LookUpRa10(Index: Integer): double;
+function LookUpRa10(Index: Integer): single;
 var
-  temp: double;
+  temp: single;
 begin
   { dezimalgeometrische Reihe Ra10 }
   temp := 1;
